Struct lumol::energy::NullPotential

pub struct NullPotential;

No-op potential.

The NullPotential always returns 0 as energy and force. It should be used to indicate that there is no potential interaction for a given set of particles.

Examples

use lumol::energy::Potential;
use lumol::energy::NullPotential;

let potential = NullPotential;
assert_eq!(potential.energy(0.1), 0.0);
assert_eq!(potential.energy(100000.0), 0.0);

assert_eq!(potential.force(0.1), 0.0);
assert_eq!(potential.force(100000.0), 0.0);Run