Struct lumol::energy::Wolf

pub struct Wolf { /* fields omitted */ }

Wolf summation for coulombic interactions.

This is a fast, direct, pairwise summation for coulombic potential [Wolf1999].

Examples

// A relatively large cutoff is needed for Wolf summation
let wolf = Wolf::new(12.0);

// Setup a system containing a NaCl pair
let mut system = System::with_cell(UnitCell::cubic(30.0));

let mut na = Particle::new("Na");
na.charge = 1.0;
na.position = Vector3D::new(0.0, 0.0, 0.0);

let mut cl = Particle::new("Cl");
cl.charge = -1.0;
cl.position = Vector3D::new(2.0, 0.0, 0.0);

system.add_molecule(Molecule::new(na));
system.add_molecule(Molecule::new(cl));

// Use Wolf summation for electrostatic interactions
system.set_coulomb_potential(Box::new(wolf));

assert_eq!(system.potential_energy(), -0.0729290269539354);

[Wolf1999]: Wolf, D. et al. J. Chem. Phys. 110, 8254 (1999).

Methods

impl Wolf

pub fn new(cutoff: f64) -> Wolf

Create a new Wolf summation, using a real-space cutoff of cutoff.