Struct lumol::input::Trajectory

pub struct Trajectory(_);

A Trajectory is a file containing one or more successives simulation steps

Methods

impl Trajectory

fn open<P: AsRef<Path>>(path: P) -> TrajectoryResult<Trajectory>

Open an existing file at path for reading.

fn create<P: AsRef<Path>>(path: P) -> TrajectoryResult<Trajectory>

Create a new file at path for writing, and overwrite any existing file.

fn read(&mut self) -> TrajectoryResult<System>

Read the next step of the trajectory

fn read_guess_bonds(&mut self) -> TrajectoryResult<System>

Read the next step of the trajectory, and guess the bonds of the resulting System.

fn write(&mut self, system: &System) -> TrajectoryResult<()>

Write the system to the trajectory.

fn as_chemfiles(&mut self) -> &mut Trajectory

Get access to the chemfiles trajectory, and the associated features