Trait lumol::input::FromTomlWithData

pub trait FromTomlWithData {
    type Data;
    fn from_toml(
        table: &Map<String, Value>, 
        data: Self::Data
    ) -> Result<Self, Error>;
}

Convert a TOML table and some additional owned data to a Rust type.

Associated Types

type Data

The type of the additional data needed.

Loading content...
Required methods

fn from_toml(
table: &Map<String, Value>,
data: Self::Data
) -> Result<Self, Error>

Do the conversion from table and data to Self.

Loading content...