Trait lumol::input::FromTomlWithRefData

pub trait FromTomlWithRefData {
    type Data;
    fn from_toml(
        table: &Map<String, Value>, 
        data: &Self::Data
    ) -> Result<Self, Error>;
}

Convert a TOML table to a Rust type using information from an additional reference.

Associated Types

type Data

The type of the additional data needed.

Loading content...
Required methods

fn from_toml(
table: &Map<String, Value>,
data: &Self::Data
) -> Result<Self, Error>

Do the conversion from table and data to Self.

Loading content...