Module lumol::potentials

Interaction potentials for energy and forces computations

Potential functions

A potential function is defined as a parametric function giving the energy and the force acting on a specific particle. These function are represented by the Potential trait, and allow to compute the energy or force norm corresponding to a scalar variable. This scalar variable will be the distance for non-bonded pair interactions; the angle for covalent angles interactions, etc.

Additional marker traits are used to specify wheter a given potential can be used for a specific interaction type. These marker traits are:

PairPotential for bonded and non-bonded two body interactions;

AnglePotential for covalent angles interactions;

DihedralPotential for covalent dihedral angles interactions.

To be able to use a specific potential in one of these contexts, the only thing to do is to implement the corresponding trait.

#[derive(Clone)]
struct OnePotential;

// OnePotential is a potential function
impl Potential for OnePotential {
    fn energy(&self, _: f64) -> f64 {1.0}
    fn force(&self, _: f64) -> f64 {0.0}
}

// It can be used for pair and dihedral potentials, but not for angles
impl PairPotential for OnePotential {}
impl DihedralPotential for OnePotential {}

Global potentials

Global potentials are potentials that have an effect on the whole system at once. They are defined by implementing the GlobalPotential trait and giving methods for energy, forces and virial contributions.

CoulombicPotential are a specific version of global potentials, that are used to compute charges-charges interactions.

Structs

CosineHarmonic
Cosine harmonic potential.

Ewald
Ewald summation of the coulombic interactions.

Harmonic
Harmonic potential.

LennardJones
Lennard-Jones potential.

NullPotential
The NullPotential always returns 0 as energy and force.

PairInteraction
A non-bonded interaction between two particle. The potential T is associated with a pair restriction. For all non-bonded pairs, the potential is computed up to a cutoff distance. An additional shifting of the potential can be used in molecular dynamics, to ensure that the energy is continuous at the cutoff distance.

RestrictionInfo
Restriction information attached to a pair of Particles in a System

TableComputation
Computation of a potential using tabulated values.

Torsion
Torsion potential.

Wolf
Wolf summation for coulombic interactions.

Enums

PairRestriction
Possible restrictions on the pair interactions.

Traits

AnglePotential
Potential that can be used for molecular angles.

BondPotential
Potential that can be used for molecular bonds

BoxCloneAngle
This is an internal implementation detail for cloning Box, where Trait is used as a trait object.

BoxCloneBond
This is an internal implementation detail for cloning Box, where Trait is used as a trait object.

BoxCloneDihedral
This is an internal implementation detail for cloning Box, where Trait is used as a trait object.

BoxClonePair
This is an internal implementation detail for cloning Box, where Trait is used as a trait object.

BoxClonePotential
This is an internal implementation detail for cloning Box, where Trait is used as a trait object.

Computation
Methods for energy and forces computation.

CoulombicPotential
Electrostatic potential solver.

DihedralPotential
Potential that can be used for molecular dihedral angles.

GlobalPotential
Potential acting on the whole system at once.

PairPotential
Potential that can be used for non-bonded two body interactions

Potential
A set of two parametric functions which takes a single scalar variable and return the corresponding energy or norm of the force.

Virial
Computation of virial contribution for a potential. The provided fucntion only apply to two-body virial contributions.