Struct lumol::potentials::Ewald

pub struct Ewald {
    // some fields omitted
}

Ewald summation of the coulombic interactions.

The Ewald summation is based on a separation of the coulombic potential U in two parts, using the trivial identity: $$U(x) = U(x) * (f(x) + 1) - U(x) * f(x)$$ where f(x) is the erf function. This leads to a separation of the conditionally convergent coulombic sum into two absolutly convergent sums: one in real space, and the other in Fourier or k-space.

For more informations about this algorithm see [FS2002].

[FS2002] Frenkel, D. & Smit, B. Understanding molecular simulation. (Academic press, 2002).

Methods

impl Ewald

fn new(rc: f64, kmax: usize) -> Ewald

Create an Ewald summation using the rc cutoff radius in real space, and kmax points in k-space (Fourier space).