Struct lumol::potentials::TableComputation

pub struct TableComputation {
    // some fields omitted
}

Computation of a potential using tabulated values.

This can be faster than direct computation for smooth potentials, but will either uses more memory or be less precise than direct computation. Values are tabulated in the [0, max) range, and a cutoff is applied after max. Energy is shifted to ensure E(max) = 0

Methods

impl TableComputation

fn new(potential: Box<PairPotential>, size: usize, max: f64) -> TableComputation

Create a new TableComputation for potential, with size points and a maximum value of max.