Trait lumol::potentials::BoxCloneBond

pub trait BoxCloneBond {
    fn box_clone_bond(&self) -> Box<BondPotential>;
}

This is an internal implementation detail for cloning Box, where Trait is used as a trait object.

Required Methods

fn box_clone_bond(&self) -> Box<BondPotential>

Get a cloned of self as a boxed trait.