Enum lumol::sim::TemperatureStrategy

pub enum TemperatureStrategy {
    None,
    Velocities,
    External(f64),
}

Possible temperature computation strategies. Different propagators needs different ways to compute the temperature: Monte-Carlo temperature is a constant of the simulation, whereas for molecular dynamics we use the instantaneous velocities.

Variants
None

No specific strategy, use whatever strategy was already in use.

Velocities

Use the instantaneous velocities to compute the temperature

External(f64)

Use a fixed external temperature