Enum lumol::sim::mc::MCDegreeOfFreedom

pub enum MCDegreeOfFreedom {
    AllMolecules,
    Molecules(BTreeSet<MoleculeHash>),
    Particles,
}

Possible degrees of freedom simulated by a given Monte Carlo move

Variants
AllMolecules

All molecules are simulated

Molecules(BTreeSet<MoleculeHash>)

All molecules with a molecule type in the BTreeSet are simulated

Particles

All the particles are simulated

Methods

impl MCDegreeOfFreedom

pub fn combine(self, other: MCDegreeOfFreedom) -> MCDegreeOfFreedom

Combine the degrees of freedom represented by this MCDegreeOfFreedom and other