Struct lumol::sim::mc::MonteCarlo

pub struct MonteCarlo { /* fields omitted */ }

Metropolis Monte Carlo propagator

Methods

impl MonteCarlo

pub fn new(temperature: f64) -> MonteCarlo

Create a new Monte Carlo propagator at temperature T.

pub fn from_rng(temperature: f64, rng: Box<RngCore + 'static>) -> MonteCarlo

Create a Monte Carlo propagator at temperature T, using the rng random number generator.

pub fn add(&mut self, mcmove: Box<MCMove + 'static>, frequency: f64)

Add the mcmove Monte Carlo move to this propagator, with frequency frequency. All calls to this function should happen before any simulation run.

Panics

If called after a simulation run.

pub fn add_move_with_acceptance(
&mut self,
mcmove: Box<MCMove + 'static>,
frequency: f64,
target_acceptance: f64
)

Add the mcmove Monte Carlo move to the propagator. frequency describes how frequent a move is called, target_acceptance is the desired acceptance ratio of the move.

Panics

If called after a simulation run. If target_acceptance is either negative or larger than one.

pub fn set_amplitude_update_frequency(&mut self, frequency: u64)

Set the number of times a move has to be called before its amplitude is updated. This value is applied to all moves.

pub fn temperature(&self) -> f64

Get the temperature of the simulation

pub fn set_temperature(&mut self, temperature: f64)

Set the temperature of the simulation