Struct lumol::sim::mc::MoveCounter

pub struct MoveCounter {
    pub total_attempted: u64,
    pub total_accepted: u64,
    pub accepted: u64,
    pub attempted: u64,
    // some fields omitted
}

This struct keeps track of the number of times a move was called and how often it was accepted.

These statistics can be used to compute a scaling factor to increase the efficiency of a move.

Example

use lumol_sim::mc::MoveCounter;

// Create a new counter with a target acceptance of 0.5 (=50 %)
let mut counter = MoveCounter::new(Some(0.5));
// Move was attempted and accepted.
counter.accept();
// Increase scaling factor since 100 % of moves where accepted.
assert_eq!(counter.compute_scaling_factor(), Some(1.2));
// Update the counters to track acceptance for move with new
// scaling factor.
counter.update();
// Directly calling for a new scaling factor returns
// `None` since there were no attempts since last update.
assert_eq!(counter.compute_scaling_factor(), None);
// The overall acceptance is still 1.0 (=100 %).
assert_eq!(counter.acceptance(), 1.0);

// The scaling factor is computed based on the acceptance
// since the last update was performed, not the overall acceptance
counter.reject();
// The scaling factor will be smaller: the acceptance
// since the last update is zero.
assert_eq!(counter.compute_scaling_factor(), Some(0.8));
assert_eq!(counter.acceptance(), 0.5);

Fields
total_attempted: u64

Count the total number of times the move was called.

total_accepted: u64

Count the total number of times the move was accepted.

accepted: u64

Count the number of times the move was accepted since the last update.

attempted: u64

Count the number of times the move was called since the last update.

Methods

impl MoveCounter

pub fn new(target_acceptance: Option<f64>) -> MoveCounter

Create a new counter for the move, initializing all counts to zero and setting the target_acceptance.

pub fn set_acceptance(&mut self, target_acceptance: Option<f64>)

Set the target acceptance for the move counter.

pub fn reject(&mut self)

Increase counters for attempt.

pub fn accept(&mut self)

Increase counters to track the number of times the move was accepted.

pub fn update(&mut self)

Reset counters for attempts and acceptance since the last update.

pub fn acceptance(&self) -> f64

Return fraction of total number of accepted over total number of attempted moves.

pub fn compute_scaling_factor(&self) -> Option<f64>

Compute a scaling factor according to the desired acceptance.