Struct lumol::sim::mc::Rotate

pub struct Rotate { /* fields omitted */ }

Monte Carlo move for rotating a rigid molecule

Methods

impl Rotate

pub fn new<H>(theta: f64, hash: H) -> Rotate where
H: Into<Option<MoleculeHash>>,

Create a new Rotate move, with maximum angular displacement of theta. This move will apply to the molecules with the given hash, or all molecules if hash is None.