Struct lumol::sim::mc::Translate

pub struct Translate { /* fields omitted */ }

Monte Carlo move for translating a molecule

Implementations

impl Translate

pub fn new<H>(delta: f64, hash: H) -> Translate where
H: Into<Option<MoleculeHash>>,

Create a new Translate move, with maximum displacement of delta. This move will apply to the molecules with the given hash, or all molecules if hash is None.