Module lumol::sim::md

Molecular dynamics related algorithms.

Each class of algorithms is represented by a trait, and each single method is a struct implementing said trait.

The main entry point is the MolecularDynamics struct, implementing the Propagator trait.

Integrator

The core of molecular dynamics simulation happens in an Integrator. The implemented integrators are the following:

Verlet: simple Verlet integrator;

VelocityVerlet: simple velocity-Verlet integrator;

LeapFrog: Leap-Frog integrator;

BerendsenBarostat: isotropic Berendsen barostat coupled to a velocity-Verlet integrator;

AnisoBerendsenBarostat anisotropic Berendsen barostat coupled to a velocity-Verlet integrator

Themostats

Thermostat are algorihtms used to fix the temperature during a simulation. The following implementation are available:

RescaleThermostat: basic themostat rescaling the velocities of all atoms. This is highly unphysical but can be usefull for equilibration;

CSVRThermostat: Canonical Sampling through Velocities Rescaling is a well-behaved thermostating algorithm generating the expected canonical ensemble distribution of states.

BerendsenThermostat: berendsen or weak-coupling thermostat;

Control

Control algorihtms group any algorithm modifying the system to enforce some type of invariant.

RemoveTranslation: remove the global translational momentum of a system;

RemoveRotation: remove the global rotational momentum of a system;

Rewrap: wrap all atoms from a system inside the unit cell;

Structs

AnisoBerendsenBarostat
Anisotropic Berendsen barostat integrator based on velocity-Verlet.

BerendsenBarostat
Berendsen barostat integrator based on velocity-Verlet.

BerendsenThermostat
Berendsen (or weak coupling) thermostat.

CSVRThermostat
A thermostat using the Canonical Sampling Through Velocities Rescaling algorithm.

LeapFrog
Leap-frog integrator.

MolecularDynamics
Molecular Dynamics propagator for the simulation.

RemoveRotation
Remove global rotation from the system

RemoveTranslation
Remove global translation from the system

RescaleThermostat
Velocity rescaling thermostat.

Rewrap
Rewrap all molecules' centers of mass to lie within the unit cell.

VelocityVerlet
Velocity-Verlet integrator.

Verlet
Verlet integrator.

Traits

Control
Trait for controlling some parameters in a system during a simulation.

Integrator
The Integrator trait define integrator interface for molecular dynamics. An integrator is an algorithm responsible for propagating the equations of motion in the system.

Thermostat
Trait for thermostat algorithms some parameters in a system during a simulation.