Struct lumol::sim::md::AnisoBerendsenBarostat

pub struct AnisoBerendsenBarostat { /* fields omitted */ }

Anisotropic Berendsen barostat integrator based on velocity-Verlet. This one neither reversible nor symplectic.

Methods

impl AnisoBerendsenBarostat

pub fn new(timestep: f64, stress: Matrix3, tau: f64) -> AnisoBerendsenBarostat

Create a new anisotropic Berendsen barostat with an integration timestep of timestep, and a target stress matrix of stress and the barostat time scale tau.

pub fn hydrostatic(
timestep: f64,
pressure: f64,
tau: f64
) -> AnisoBerendsenBarostat

Create a new anisotropic Berendsen barostat with an integration timestep of timestep, using an hydrostatic stress matrix corresponding to the pressure pressure and the barostat time scale tau.