Trait lumol::sim::md::Control

pub trait Control {
    fn control(&mut self, system: &mut System);

    fn setup(&mut self, _: &System) { ... }
    fn finish(&mut self, _: &System) { ... }
}

Trait for controling some parameters in a system during a simulation.

Required Methods

fn control(&mut self, system: &mut System)

Do your job, control algorithm!

Provided Methods

fn setup(&mut self, _: &System)

Function called once at the beggining of the simulation, which allow for some setup of the control algorithm if needed.

fn finish(&mut self, _: &System)

Function called once at the end of the simulation.