Trait lumol::sim::md::Integrator

pub trait Integrator {
    fn integrate(&mut self, system: &mut System);

    fn setup(&mut self, &System) { ... }
}

The Integrator trait define integrator interface for molecular dynamics. An integrator is an algorithm responsible for propagating the equations of motion in the system.

Required methods

fn integrate(&mut self, system: &mut System)

Integrate the equations of motion. This is called at every step of the simulation.

Loading content...
Provided methods

fn setup(&mut self, &System)

Setup the integrator. This function is called once by every simulation run.

Loading content...