Struct lumol::sim::min::Minimization

pub struct Minimization { /* fields omitted */ }

Minimization propagator for simulations.

The minimization stops when the energy difference between the previous and the current step is lower than the energy criterion, or when the maximal squared norm of the atomic force is lower than the force criterion.

Methods

impl Minimization

pub fn new(
minimizer: Box<Minimizer + 'static>,
tolerance: Tolerance
) -> Minimization

Create a new Minimization using the given minimizer and specific energy and force tolerance.

pub fn converged(&self) -> bool

Check if the minimization has converged.