Struct lumol::sim::min::SteepestDescent

pub struct SteepestDescent { /* fields omitted */ }

Steepest gradient descent for energy minimization.

This algorithm is very rough, and will not converge in all the situations. However it is easy to use, and simple enough to be implemented quickly.

Methods

impl SteepestDescent

fn new() -> SteepestDescent

Create a SteepestDescent with sensible default values for energy and force convergence criteria.

The default for force criterion is 1e-5 kJ/mol/Å^2, and 1e-5 kJ/mol/Å^2 for the energy criterion.

fn with_criteria(force: f64, energy: f64) -> SteepestDescent

Create a new SteepestDescent with the given force and energy convergence criterion.

fn converged(&self) -> bool

Has the minimization converged so far ?