Trait lumol::sim::min::Minimizer

pub trait Minimizer {
    fn minimize(&mut self, system: &mut System) -> Tolerance;

    fn degrees_of_freedom(&self, system: &System) -> DegreesOfFreedom;

    fn setup(&mut self, &System) { ... }
}

The Minimizer trait define minimization interface.

A minimizer is an algorithm responsible for finding new configurations of lower energy.

Required methods

fn minimize(&mut self, system: &mut System) -> Tolerance

Find a new configuration of lower energy, and return the corresponding values for energy and forces.

fn degrees_of_freedom(&self, system: &System) -> DegreesOfFreedom

Get the number of degrees of freedom simulated by this minimizer

This function is called once at thr beginning of the simulation

Loading content...
Provided methods

fn setup(&mut self, &System)

Setup the minimizer. This function is called once at the begining of every simulation run.

Loading content...