Struct lumol::sim::Alternator

pub struct Alternator<T> { /* fields omitted */ }

Helper struct that can wrap an algorithm to make it run only a fraction of the times it is called.

Example

use lumol::sim::Alternator;

struct HelloWorld;

trait SaySomething {
    fn say_something(&mut self) -> &'static str;
}

impl SaySomething for HelloWorld {
    fn say_something(&mut self) -> &'static str {
        "Hello world!"
    }
}

impl<T> SaySomething for Alternator<T> where T: SaySomething {
    fn say_something(&mut self) -> &'static str {
        if self.can_run() {
            self.as_mut().say_something()
        } else {
            ""
        }
    }
}

let mut alternator = Alternator::new(3, HelloWorld);

assert_eq!(alternator.say_something(), "");
assert_eq!(alternator.say_something(), "");
assert_eq!(alternator.say_something(), "Hello world!");
assert_eq!(alternator.say_something(), "");
assert_eq!(alternator.say_something(), "");
assert_eq!(alternator.say_something(), "Hello world!");

Methods

impl<T> Alternator<T>

fn new(every: u64, base: T) -> Alternator<T>

Wrap the algorithm base to call it only every every time.

fn can_run(&mut self) -> bool

Check if this is the appropriate time to run the algorithm.