Struct lumol::sim::Simulation

pub struct Simulation { /* fields omitted */ }

The Simulation struct holds all the needed algorithms for running the simulation. It should be use together with a System to perform the simulation.

Methods

impl Simulation

pub fn new(propagator: Box<Propagator + 'static>) -> Simulation

Create a new Simulation from a Propagator.

pub fn run(&mut self, system: &mut System, nsteps: usize)

Run the simulation on System for nsteps steps.

pub fn add_output(&mut self, output: Box<Output + 'static>)

Add a new Output algorithm in the outputs list

pub fn add_output_with_frequency(
&mut self,
output: Box<Output + 'static>,
frequency: u64
)

Add a new Output algorithm in the outputs list, which will be used at the given frequency. The output will be used every time the system step matches this frequency.