Trait lumol::sim::Propagator

pub trait Propagator {
    fn temperature_strategy(&self) -> TemperatureStrategy;

    fn degrees_of_freedom(&self, system: &System) -> DegreesOfFreedom;

    fn propagate(&mut self, system: &mut System);

    fn setup(&mut self, &System) { ... }

    fn finish(&mut self, &System) { ... }
}

The propagator trait is the main algorithm of a simulation, i.e. the one which update the system. The main function here is propagate, which should propagate the simulation for one step.

Required methods

fn temperature_strategy(&self) -> TemperatureStrategy

Get the temperature computation strategy for this propagator.

This function is called once at thr beginning of the simulation

fn degrees_of_freedom(&self, system: &System) -> DegreesOfFreedom

Get the number of degrees of freedom simulated by this propagator

This function is called once at thr beginning of the simulation

fn propagate(&mut self, system: &mut System)

Propagate the system for one simulation step.

Loading content...
Provided methods

fn setup(&mut self, &System)

Setup code, preparing all the meta-information needed about the simulation.

fn finish(&mut self, &System)

Finish the simulation, and maybe output some information about it

Loading content...