Struct lumol::simulation::mc::MonteCarlo

pub struct MonteCarlo {
    // some fields omitted
}

Metropolis Monte-Carlo propagator

Methods

impl MonteCarlo

fn new(temperature: f64) -> MonteCarlo

Create a new Monte-Carlo propagator at temperature T.

fn from_rng(temperature: f64, rng: Box<Rng>) -> MonteCarlo

Create a Monte-Carlo propagator at temperature T, using the rng random number generator.

fn add(&mut self, mcmove: Box<MCMove>, frequency: f64)

Add a the mcmove Monte-Carlo move to this propagator, with frequency frequency. All calls to this function should happen before any simulation run.

Panics

If called after a simulation run.

fn temperature(&self) -> f64

Get the temperature of the simulation

fn set_temperature(&mut self, temperature: f64)

Set the temperature of the simulation