Struct lumol::simulation::md::RescaleThermostat

pub struct RescaleThermostat {
    // some fields omitted
}

Velocity rescaling thermostat.

The velocity rescale algorithm controls the temperature by rescaling all the velocities when the temperature differs exceedingly from the desired temperature. A tolerance parameter prevent this algorithm from running too often: if tolerance is 10K and the target temperature is 300K, the algorithm will only run if the instant temperature is below 290K or above 310K.

Methods

impl RescaleThermostat

fn new(temperature: f64) -> RescaleThermostat

Create a new RescaleThermostat acting at temperature temperature, with a tolerance of 5% * temperature.

fn with_tolerance(temperature: f64, tol: f64) -> RescaleThermostat

Create a new RescaleThermostat acting at temperature T, with a tolerance of tol. For rescaling all the steps, use tol = 0.