Trait lumol::sys::compute::Compute

pub trait Compute {
    type Output;
    fn compute(&self, system: &System) -> Self::Output;
}

The Compute trait allow to compute properties of a system, without modifying this system. The Output type is the type of the computed property.

Associated Types

type Output

The data type of the property

Loading content...
Required methods

fn compute(&self, system: &System) -> Self::Output

Compute the property

Loading content...