Struct lumol::sys::EnergyEvaluator

pub struct EnergyEvaluator<'a> { /* fields omitted */ }

An helper struct to evaluate energy components of a system.

Methods

impl<'a> EnergyEvaluator<'a>

pub fn new(system: &'a System) -> EnergyEvaluator<'a>

Create a new EnergyEvaluator acting on the given system.

impl<'a> EnergyEvaluator<'a>

pub fn pair(&self, path: BondPath, r: f64, i: usize, j: usize) -> f64

Compute the energy associated with the pair of particles i, j at distance r

pub fn pairs(&self) -> f64

Compute the energy of all the pairs in the system

pub fn pairs_tail(&self) -> f64

Compute the energy due to long range corrections for the pairs

pub fn bond(&self, r: f64, i: usize, j: usize) -> f64

Compute the energy associated with the bonded particles i, j at distance r

pub fn bonds(&self) -> f64

Compute the energy of all the bonds in the system

pub fn angle(&self, theta: f64, i: usize, j: usize, k: usize) -> f64

Compute the energy associated with the angle i, j, k at angle theta

pub fn angles(&self) -> f64

Compute the energy of all the angles in the system

pub fn dihedral(&self, phi: f64, i: usize, j: usize, k: usize, m: usize) -> f64

Compute the energy associated with the dihedral angle i, j, k, m at angle phi

pub fn dihedrals(&self) -> f64

Compute the energy of all the dihedral angles in the system

pub fn coulomb(&self) -> f64

Compute the energy of the electrostatic interactions

pub fn global(&self) -> f64

Compute the energy of the global potentials