Struct lumol::sys::Molecule

pub struct Molecule { /* fields omitted */ }

A Molecule associate some particles bonded together.

Molecule implement Deref to a Bonding struct, to give read access to all the bonds. It does not implement DerefMut, adding new bonds should be done through Molecule::add_bond().

Methods

impl Molecule

pub fn new(particle: Particle) -> Molecule

Create a new Molecule containing a single particle

pub fn as_ref(&self) -> MoleculeRef

Borrow self as a MoleculeRef.

pub fn as_mut(&mut self) -> MoleculeRefMut

Mutablely borrow self as a MoleculeRefMut.

pub fn particles(&self) -> ParticleSlice

Get access to the particles in this molecule

pub fn particles_mut(&mut self) -> ParticleSliceMut

Get mutable access to the particles in this molecule

pub fn add_particle_bonded_to(&mut self, other: usize, particle: Particle)

Add a new particle in this molecule, bonded to an other particle in the molecule.

pub fn add_bond(&mut self, i: usize, j: usize)

Add bond between particles at indexes i and j in this molecule.

Panics

If i or j are not in this molecule.

impl<'a> Molecule

pub fn center_of_mass(&self) -> Vector3D

Return the center-of-mass of a molecule

Warning

This function does not check for the particles' positions' nearest images. To use this function properly, make sure that all particles of the molecule are adjacent.

pub fn hash(&self) -> MoleculeHash

Get a hash of this molecule. This is a hash of the particles names (in order), and the set of bonds in the molecule. This means that two molecules will have the same type if and only if they contains the same atoms and the same bonds, in the same order.

impl<'a> Molecule

pub fn wrap(&mut self, cell: &UnitCell)

Move all particles of a molecule such that the molecules center-of-mass position resides inside the simulation cell.

Note

If the CellShape is Infinite there are no changes to the positions.