Struct lumol::sys::MoleculeRefMut

pub struct MoleculeRefMut<'a> { /* fields omitted */ }

An analog to [&mut Molecule] using particles stored elsewhere (in a system or an Molecule).

MoleculeRefMut implement Deref to a [Bonding] struct, to give read access to all the bonds. It does not implement DerefMut, adding new bonds should be done through Molecule::add_bond() or Configuration::add_bond().

Methods

impl<'a> MoleculeRefMut<'a>

pub fn new(
bonding: &'a Bonding,
particles: ParticleSliceMut<'a>
) -> MoleculeRefMut<'a>

Create a new MoleculeRefMut associating the given bonding and particles.

Panics

If the bonding and the particles do not containe the same number of particles.

pub fn as_ref(&self) -> MoleculeRef

Borrow self as a MoleculeRef.

pub fn particles(&self) -> ParticleSlice

Get access to the particles in this molecule

pub fn particles_mut(&mut self) -> ParticleSliceMut

Get mutable access to the particles in this molecule

pub fn to_owned(&self) -> Molecule

Copies self into a new Molecule

impl<'a> MoleculeRefMut<'a>

pub fn center_of_mass(&self) -> Vector3D

Return the center-of-mass of a molecule

Warning

This function does not check for the particles' positions' nearest images. To use this function properly, make sure that all particles of the molecule are adjacent.

pub fn hash(&self) -> MoleculeHash

Get a hash of this molecule. This is a hash of the particles names (in order), and the set of bonds in the molecule. This means that two molecules will have the same type if and only if they contains the same atoms and the same bonds, in the same order.

impl<'a> MoleculeRefMut<'a>

pub fn wrap(&mut self, cell: &UnitCell)

Move all particles of a molecule such that the molecules center-of-mass position resides inside the simulation cell.

Note

If the CellShape is Infinite there are no changes to the positions.