Struct lumol::sys::ParticleSliceMut

pub struct ParticleSliceMut<'a> {
    pub name: &'a mut [String],
    pub kind: &'a mut [ParticleKind],
    pub charge: &'a mut [f64],
    pub mass: &'a mut [f64],
    pub position: &'a mut [Vector3D],
    pub velocity: &'a mut [Vector3D],
}

A mutable slice of Particle inside a ParticleVec .

Fields
name: &'a mut [String]

A mutable slice of name from a ParticleVec

kind: &'a mut [ParticleKind]

A mutable slice of kind from a ParticleVec

charge: &'a mut [f64]

A mutable slice of charge from a ParticleVec

mass: &'a mut [f64]

A mutable slice of mass from a ParticleVec

position: &'a mut [Vector3D]

A mutable slice of position from a ParticleVec

velocity: &'a mut [Vector3D]

A mutable slice of velocity from a ParticleVec

Methods

impl<'a> ParticleSliceMut<'a>

pub fn as_ref(&self) -> ParticleSlice

Convert a ParticleSliceMut to a ParticleSlice in order to be able to use the methods on the non mutable version of the slices.

pub fn len(&self) -> usize

Similar to [Particle] ::len(), the length of all fields should be the same.

pub fn is_empty(&self) -> bool

Similar to [Particle] ::is_empty(), the length of all fields should be the same.

pub fn first_mut(&mut self) -> Option<ParticleRefMut>

Similar to [Particle] ::first_mut().

pub fn split_first_mut(&mut self) -> Option<(ParticleRefMut, ParticleSliceMut)>

Similar to [Particle] ::split_first_mut().

pub fn last_mut(&mut self) -> Option<ParticleRefMut>

Similar to [Particle] ::last_mut().

pub fn split_last_mut(&mut self) -> Option<(ParticleRefMut, ParticleSliceMut)>

Similar to [Particle] ::last_mut().

pub fn split_at_mut(
&mut self,
mid: usize
) -> (ParticleSliceMut, ParticleSliceMut)

Similar to [Particle] ::split_at_mut().

pub fn swap(&mut self, a: usize, b: usize)

Similar to [Particle] ::swap().

pub fn get(&self, i: usize) -> Option<ParticleRef>

Similar to [Particle] ::get().

pub unsafe fn get_unchecked(&self, i: usize) -> ParticleRef

Similar to [Particle] ::get_unchecked().

pub fn get_mut(&mut self, i: usize) -> Option<ParticleRefMut>

Similar to [Particle] ::get_mut().

pub unsafe fn get_unchecked_mut(&mut self, i: usize) -> ParticleRefMut

Similar to [Particle] ::get_unchecked_mut().

pub fn as_ptr(&self) -> ParticlePtr

Similar to [Particle] ::as_ptr().

pub fn as_mut_ptr(&mut self) -> ParticlePtrMut

Similar to [Particle] ::as_mut_ptr().

pub unsafe fn from_raw_parts_mut<'b>(
data: ParticlePtrMut,
len: usize
) -> ParticleSliceMut<'b>

Similar to std::slice::from_raw_parts_mut().

impl<'a> ParticleSliceMut<'a>

pub fn to_vec(&self) -> ParticleVec

Similar to [Particle] ::to_vec().

impl<'a> ParticleSliceMut<'a>

pub fn iter(&mut self) -> Iter

Get an iterator over the ParticleRef in this vector

pub fn iter_mut(&mut self) -> IterMut

Get a mutable iterator over the ParticleRefMut in this vector