Struct lumol::sys::TrajectoryBuilder

pub struct TrajectoryBuilder<'a> { /* fields omitted */ }

A Trajectory builder, to set some options before opening a trajectory.

Examples

let trajectory = TrajectoryBuilder::new()
                                   .mode(OpenMode::Read)
                                   .open("file.xyz")
                                   .unwrap();

Methods

impl<'a> TrajectoryBuilder<'a>

fn new() -> TrajectoryBuilder<'a>

Create a new builder in read mode and with automatic format detection.

Examples

let trajectory = TrajectoryBuilder::new()
                                   .open("file.xyz")
                                   .unwrap();

fn format(self, format: &'a str) -> TrajectoryBuilder<'a>

Use a specific format when opening a file. See the chemfiles documentation for a format list.

Examples

let trajectory = TrajectoryBuilder::new()
                                   .format("PDB")
                                   .open("file.mol")
                                   .unwrap();

fn mode(self, mode: OpenMode) -> TrajectoryBuilder<'a>

Use a specific mode when opening a file.

Examples

let trajectory = TrajectoryBuilder::new()
                                   .mode(OpenMode::Write)
                                   .open("file.nc")
                                   .unwrap();

fn open<P: AsRef<Path>>(self, path: P) -> Result<Trajectory, TrajectoryError>

Open the trajectory at the given path.

Examples

let trajectory = TrajectoryBuilder::new()
                                   .open("file.nc")
                                   .unwrap();