Struct lumol::system::compute::StressAtTemperature

pub struct StressAtTemperature {
    pub temperature: f64,
}

Compute the stress tensor of the system from the virial equation, at the given temperature. The stress tensor is defined by $$ \sigma = \sigma = \frac{1}{V} (\sum_i m_i v_i \otimes v_i + \sum_i \sum_{j > i} \vec r_{ij} \otimes \vec f_{ij}) $$ but here the kinetic energy term is replaced by it average at the given temperature.

Fields
temperature: f64

Temperature for the stress tensor computation