Struct lumol::system::EnergyEvaluator

pub struct EnergyEvaluator<'a> {
    // some fields omitted
}

An helper struct to evaluate energy components of a system.

Methods

impl<'a> EnergyEvaluator<'a>

fn new(system: &'a System) -> EnergyEvaluator<'a>

Create a new energy evaluator acting on the given system.

impl<'a> EnergyEvaluator<'a>

fn pair(&self, r: f64, i: usize, j: usize) -> f64

Compute the energy associated with the pair of particles i, j at distance r

fn pairs(&self) -> f64

Compute the energy of all the pairs in the system

fn bond(&self, r: f64, i: usize, j: usize) -> f64

Compute the energy associated with the bonded particles i, j at distance r

fn bonds(&self) -> f64

Compute the energy of all the bonds in the system

fn angle(&self, theta: f64, i: usize, j: usize, k: usize) -> f64

Compute the energy associated with the angle i, j, k at angle theta

fn angles(&self) -> f64

Compute the energy of all the angles in the system

fn dihedral(&self, phi: f64, i: usize, j: usize, k: usize, m: usize) -> f64

Compute the energy associated with the dihedral angle i, j, k, m at angle phi

fn dihedrals(&self) -> f64

Compute the energy of all the dihedral angles in the system

fn coulomb(&self) -> f64

Compute the energy of the electrostatic interactions

fn global(&self) -> f64

Compute the energy of the global potentials