Struct lumol::system::Particle

pub struct Particle {
    pub kind: ParticleKind,
    pub mass: f64,
    pub charge: f64,
    pub position: Vector3D,
    pub velocity: Vector3D,
    // some fields omitted
}

The Particle type hold basic data about a particle in the system. It is self contained, so that it will be easy to send data between parrallels processes.

Fields
kind: ParticleKind

Particle kind, an index for potentials lookup

mass: f64

Particle mass

charge: f64

Particle charge

position: Vector3D

Particle positions

velocity: Vector3D

Particle velocity, if needed

Methods

impl Particle

fn new<S: Into<String>>(name: S) -> Particle

Create a new Particle from a name

fn name(&self) -> &str

Get the particle name

fn set_name<'a, S>(&mut self, name: S) where S: Into<&'a str>

Set the particle name to name