Struct lumol::system::System

pub struct System {
    // some fields omitted
}

The System type hold all the data about a simulated system.

This data contains:

an unit cell, containing the system;

a list of particles in the system;

a list of molecules in the system;

a list of interactions, associating particles kinds and potentials

a hash map associating particles names and particles kinds.

In the implementation, the particles contained in a molecule are guaranted to be contiguous in memory. This allow for faster access when iterating over molecules, and easier molecule removal in the system.

Methods

impl System

fn new() -> System

Create a new empty System

fn from_cell(cell: UnitCell) -> System

Create an empty system with a specific UnitCell

fn step(&self) -> u64

Get the current step of the system

fn increment_step(&mut self)

Increment the system step

impl System

Topology and particles related functions

fn moltype(&self, molid: usize) -> u64

Get the type of the molecule at index molid. This type is a hash of the atoms names, and the set of bonds in the molecule. This means that two molecules will have the same type if and only if they contains the same atoms and the same bonds, in the same order.

fn molecules_with_moltype(&self, moltype: u64) -> Vec<usize>

Get a list of molecules with moltype molecule type.

fn are_in_same_molecule(&self, i: usize, j: usize) -> bool

Check if the particles at indexes i and j are in the same molecule

fn molecules(&self) -> &[Molecule]

Get the list of molecules in the system.

fn molecule(&self, id: usize) -> &Molecule

Get the molecule at index id

fn molid(&self, i: usize) -> usize

Get the index of the molecule containing the particle i

fn bond_distance(&self, i: usize, j: usize) -> i8

Get the length of the shortest bond path to go from the particle i to the particle j. If the particles are not in the same molecule, the length is -1. Else, this length is 0 if i == j, 1 if there is a bond between i and j, etc.

fn remove_molecule(&mut self, molid: usize)

Remove the molecule at index i

fn add_bond(&mut self, particle_i: usize, particle_j: usize) -> Permutations

Add a bond between the particles at indexes i and j. The particles should have been added to the system before calling this.

Warning

If the bond is between two particles which are not in the same molecule, the two molecules are merged together by deplacing particles in the particles list, and thus invalidate any previously stored index. In particular, any bond, angle, dihedral or molecule is invalidated.

This function will return the list of atomic permutations that where applied in order to ensure that molecules are contiguous in memory.

fn remove_particle(&mut self, i: usize)

Removes particle at index i and any associated bonds, angle or dihedral

fn add_particle(&mut self, p: Particle)

Insert a particle at the end of the internal list

fn size(&self) -> usize

Get the number of particles in this system

fn iter(&self) -> Iter<Particle>

Get an iterator over the Particle in this system

fn iter_mut(&mut self) -> IterMut<Particle>

Get a mutable iterator over the Particle in this system

impl System

Potentials related functions

fn energy_evaluator(&self) -> EnergyEvaluator

Get an helper struct to evaluate the energy of this system.

fn interactions(&self) -> &Interactions

Access the interactions for this system

fn interactions_mut(&mut self) -> &mut Interactions

Access the interactions for this system in a mutable way

fn pair_potentials(&self, i: usize, j: usize) -> &[PairInteraction]

Get the list of pair potential acting between the particles at indexes i and j.

fn bond_potentials(&self, i: usize, j: usize) -> &[Box<BondPotential>]

Get the list of bonded potential acting between the particles at indexes i and j.

fn angle_potentials(&self, i: usize, j: usize, k: usize) -> &[Box<AnglePotential>]

Get the list of angle interaction acting between the particles at indexes i, j and k.

fn dihedral_potentials(&self, i: usize, j: usize, k: usize, m: usize) -> &[Box<DihedralPotential>]

Get the list of dihedral angles interaction acting between the particles at indexes i, j, k and m.

impl System

UnitCell related functions

fn cell(&self) -> &UnitCell

Get a reference to the system unit cell

fn cell_mut(&mut self) -> &mut UnitCell

Get a mutable reference to the system unit cell

fn set_cell(&mut self, cell: UnitCell)

Set the system unit cell

fn distance(&self, i: usize, j: usize) -> f64

Get the distance between the particles at indexes i and j

fn nearest_image(&self, i: usize, j: usize) -> Vector3D

Get the vector between the nearest image of particle j with respect to particle i.

fn angle(&self, i: usize, j: usize, k: usize) -> f64

Get the angle between the particles i, j and k

fn angle_and_derivatives(&self, i: usize, j: usize, k: usize) -> (f64, Vector3D, Vector3D, Vector3D)

Get the angle and the derivatives of the angle between the particles i, j and k

fn dihedral(&self, i: usize, j: usize, k: usize, m: usize) -> f64

Get the dihedral angle between the particles i, j, k and m

fn dihedral_and_derivatives(&self, i: usize, j: usize, k: usize, m: usize) -> (f64, Vector3D, Vector3D, Vector3D, Vector3D)

Get the dihedral angle and the derivatives of the dihedral angle between the particles i, j, k and m

impl System

Functions to get pysical properties of a system.

fn kinetic_energy(&self) -> f64

Get the kinetic energy of the system.

fn potential_energy(&self) -> f64

Get the potential energy of the system.

fn total_energy(&self) -> f64

Get the total energy of the system.

fn external_temperature(&mut self, temperature: Option<f64>)

Use an external temperature for all the system properties. Calling this with Some(temperature) will replace all the computation of the temperature from the velocities with the given values. Calling it with None will use the velocities.

The default is to use the velocities unless this function is called.

fn temperature(&self) -> f64

Get the temperature of the system.

fn volume(&self) -> f64

Get the volume of the system.

fn virial(&self) -> Matrix3

Get the tensorial virial of the system.

fn pressure(&self) -> f64

Get the pressure of the system from the virial equation, at the system instananeous temperature.

fn stress(&self) -> Matrix3

Get the stress tensor of the system from the virial equation.

fn forces(&self) -> Vec<Vector3D>

Get the forces acting on all the particles in the system