Struct lumol::system::UnitCell

pub struct UnitCell {
    // some fields omitted
}

An UnitCell defines the system physical boundaries.

The shape of the cell can be any of the CellShape, and will influe on how periodic boundary conditions are applied.

Methods

impl UnitCell

fn new() -> UnitCell

Create an infinite unit cell

fn ortho(a: f64, b: f64, c: f64) -> UnitCell

Create an orthorombic unit cell, with side lengths a, b, c.

fn cubic(length: f64) -> UnitCell

Create a cubic unit cell, with side lengths length, length, length.

fn triclinic(a: f64, b: f64, c: f64, alpha: f64, beta: f64, gamma: f64) -> UnitCell

Create a triclinic unit cell, with side lengths a, b, c and angles alpha, beta, gamma.

fn shape(&self) -> CellShape

Get the cell shape

fn vect_a(&self) -> Vector3D

Get the first vector of the cell

fn a(&self) -> f64

Get the first length of the cell (i.e. the norm of the first vector of the cell)

fn vect_b(&self) -> Vector3D

Get the second vector of the cell

fn b(&self) -> f64

Get the second length of the cell (i.e. the norm of the second vector of the cell)

fn vect_c(&self) -> Vector3D

Get the third vector of the cell

fn c(&self) -> f64

Get the third length of the cell (i.e. the norm of the third vector of the cell)

fn lengths(&self) -> (f64, f64, f64)

Get the distances between faces of the unit cell

fn alpha(&self) -> f64

Get the first angle of the cell

fn beta(&self) -> f64

Get the second angle of the cell

fn gamma(&self) -> f64

Get the third angle of the cell

fn volume(&self) -> f64

Get the volume angle of the cell

fn scale_mut(&mut self, factor: Matrix3)

Scale this unit cell in-place by multiplying the cell matrix by factor.

fn scale(&self, s: Matrix3) -> UnitCell

Scale this unit cell by multiplying the cell matrix by s, and return a new scaled unit cell

fn reciprocal_vectors(&self) -> (Vector3D, Vector3D, Vector3D)

Get the reciprocal vectors of this unit cell

impl UnitCell

Geometric operations using periodic boundary conditions

fn wrap_vector(&self, vect: &mut Vector3D)

Wrap a vector in the unit cell, obeying the periodic boundary conditions. For a cubic cell of side lenght L, this produce a vector with all components in [0, L).

fn vector_image(&self, vect: &mut Vector3D)

Find the image of a vector in the unit cell, obeying the periodic boundary conditions. For a cubic cell of side lenght L, this produce a vector with all components in [-L/2, L/2).

fn fractional(&self, vect: &Vector3D) -> Vector3D

Get the fractional representation of the v vector in this cell

fn cartesian(&self, frac: &Vector3D) -> Vector3D

Get the cartesian representation of the fractional v vector in this cell

fn distance(&self, u: &Vector3D, v: &Vector3D) -> f64

Periodic boundary conditions distance between the point u and the point v

fn angle(&self, a: &Vector3D, b: &Vector3D, c: &Vector3D) -> f64

Get the angle formed by the points at a, b and c using periodic boundary conditions.

fn angle_and_derivatives(&self, a: &Vector3D, b: &Vector3D, c: &Vector3D) -> (f64, Vector3D, Vector3D, Vector3D)

Get the angle formed by the points at a, b and c using periodic boundary conditions and its derivatives.

fn dihedral(&self, a: &Vector3D, b: &Vector3D, c: &Vector3D, d: &Vector3D) -> f64

Get the dihedral angle formed by the points at a, b, c, d using periodic boundary conditions.

fn dihedral_and_derivatives(&self, a: &Vector3D, b: &Vector3D, c: &Vector3D, d: &Vector3D) -> (f64, Vector3D, Vector3D, Vector3D, Vector3D)

Get the dihedral angle and and its derivatives defined by the points at a, b, c and d using periodic boundary conditions.