Trait lumol::DihedralPotential

pub trait DihedralPotential: Potential + BoxCloneDihedral { }

Marker trait for potentials that can be used for molecular dihedral angles.

Example

use lumol_core::energy::{Potential, DihedralPotential};

// A no-op potential
#[derive(Clone)]
struct Null;

impl Potential for Null {
    fn energy(&self, x: f64) -> f64 {0.0}
    fn force(&self, x: f64) -> f64 {0.0}
}

// Now we can use the Null potential for dihedral angles
impl DihedralPotential for Null {}