Struct lumol::types::Array2

pub struct Array2<T>(_);

Two dimensional tensors, based on ndarray.

Most of the methods are simply forwarded to ndarray, so also look the doc for this crate. This array type mainly supports indexing using tuples as indices and is though as storage backend for multi-dimensional data.

let mut a = Array2::zeros((3, 5));

assert_eq!(a[(0, 4)], 0.0);

a[(0, 4)] = 7.0;
assert_eq!(a[(0, 4)], 7.0);

Implementations

impl<T> Array2<T> where
T: Clone + Zero,

pub fn zeros(size: (usize, usize)) -> Array2<T>

Create a new Array2 of the specified size filled with the Zero::zero return value.

Examples

let a: Array2<f64> = Array2::zeros((8, 5));
assert_eq!(a[(6, 2)], 0.0);

pub fn resize_if_different(&mut self, size: (usize, usize))

Resize the array if the current size is not size, and fill the new array with zeros.

Examples

let mut a = Array2::zeros((8, 5));

a[(3, 3)] = 42.0;

// This does nothing
a.resize_if_different((8, 5));
assert_eq!(a[(3, 3)], 42.0);

// This allocates a new array
a.resize_if_different((8, 9));
assert_eq!(a[(3, 3)], 0.0);

impl<T> Array2<T> where
T: Default,

pub fn default(size: (usize, usize)) -> Array2<T>

Create a new Array2 of the specified size filled with the Default::default return value.

Examples

let a: Array2<f64> = Array2::zeros((8, 5));
let b: Array2<f64> = Array2::default((8, 5));

assert_eq!(a, b);