Module lumol::units

This module allow to convert from and to the internal unit system.

Internal units are:

Angstrom (A) for distances;

femtosecond (fs) for time;

Unified atomic mass unit (u or Da) for mass;

Kelvin (K) for temperature;

Number of particles for quantity of matter;

radian (rad) for angles;

Other units are derived from these primitives units. For examples, the internal unit for energy is 1e-4 kJ/mol.

Structs

CONVERSION_FACTORS
A map of conversion factors from various units to lumol internal units

Enums

ParseError
Possible error causes when parsing an unit string.

Functions

from
Convert the numeric value val from the unit unit to the internal unit.

from_str
Parse the string val and convert it to the corresponding internal unit

to
Convert the numeric value val (in internal units) to the unit unit.