Crate lumol

Lumol is a classical molecular simulation engine that provides a solid base for developing new algorithms and methods.

Using Lumol, you can customize the behavior of all the algorithms in a simulation (from force fields to barostats and Monte-Carlo moves).

Lumol goals are to be:

Easy to extend: the code is modular, object-oriented, well documented, well tested, open-source and readable;

Easy to use: the user interface is nice, with human-oriented input files;

Stable: it will never crash on a good input, and provides helpful error messages.

Reexports

pub use types::*;

pub use potentials::*;

pub use system::*;

pub use simulation::*;

pub use chfl::*;

Modules

chfl
Chemfiles adaptators for Lumol.

constants
Usefull physical constants, expressed in the internal unit system.

potentials
Interaction potentials for energy and forces computations

simulation
Types and traits for representing simulation algorithms

system
The system module provide a way to store data about a simulated system. These systems are represented by an System instance, made of a list of Particle, an enclosing UnitCell and some interactions.

types
Some basic types used in all the other modules

units
This module allow to convert from and to the internal unit system.

Structs

Logger
Logger with capacity to write to the standard output stream, the standard error stream or a file.

Enums

LogLevel
An enum representing the available verbosity levels of the logging framework