Crate lumol

Lumol is a classical molecular simulation engine that provides a solid base for developing new algorithms and methods.

Using Lumol, you can customize the behavior of all the algorithms in a simulation (from force fields to barostats and Monte Carlo moves).

Lumol goals are to be:

Easy to extend: the code is modular, object-oriented, well documented, well tested, open-source and readable;

Easy to use: the user interface is nice, with human-oriented input files;

Stable: it will never crash on a good input, and provides helpful error messages.

Modules

consts
Useful physical constants, expressed in the internal unit system.

energy
Interaction potentials for energy and forces computations

input
This module provide a way to build a Lumol simulation using input files.

sim
Type and algorithms for simulations

sys
Representations of a simulated system

types
This module provides complexe numbers; 3D vectors and matrix; and multidimensional arrays for use in all other modules.

units
This module allow to convert from and to the internal unit system.

Statics

VERSION
The full version of the crate, containing git state if available