Struct lumol::energy::Ewald

pub struct Ewald { /* fields omitted */ }

Ewald summation for coulombic interactions.

The Ewald summation is based on a separation of the coulombic potential U in two parts, using the trivial identity: U(x) = U(x) * (f(x) + 1) - U(x) * f(x) where f is the erf function. This leads to a separation of the conditionally convergent coulombic sum into two absolutely convergent sums: one in real space, and the other in Fourier or k-space. For more information about this algorithm see [FS2002].

Examples

use lumol::energy::{Ewald, SharedEwald};
use lumol::units;

let ewald = SharedEwald::new(
    Ewald::new(/* cutoff */ 12.0, /* kmax */ 7)
);

use lumol::sys::System;
use lumol::sys::Particle;
use lumol::sys::UnitCell;
use lumol::types::Vector3D;

// Setup a system containing a NaCl pair
let mut system = System::with_cell(UnitCell::cubic(10.0));

let mut na = Particle::new("Na");
na.charge = 1.0;
na.position = Vector3D::new(0.0, 0.0, 0.0);

let mut cl = Particle::new("Cl");
cl.charge = -1.0;
cl.position = Vector3D::new(2.0, 0.0, 0.0);

system.add_particle(na);
system.add_particle(cl);

// Use Ewald summation for electrostatic interactions
system.set_coulomb_potential(Box::new(ewald));

assert_eq!(system.potential_energy(), -0.07042996180522723);

[FS2002] Frenkel, D. & Smith, B. Understanding molecular simulation. (Academic press, 2002).

Methods

impl Ewald

fn new(cutoff: f64, kmax: usize) -> Ewald

Create an Ewald summation using the given cutoff radius in real space, and kmax points in k-space (Fourier space).

fn set_alpha(&mut self, alpha: f64)

Set the value of the alpha parameter for ewald computation. The default is to use alpha = 3 * π / (4 * rc).