Struct lumol::energy::TableComputation

pub struct TableComputation { /* fields omitted */ }

Computation of a potential using tabulated values.

This can be faster than direct computation for smooth potentials, but will uses more memory and be less precise than direct computation. Values are tabulated in the [0, max) range, and a cutoff is applied after max. Energy is shifted to ensure E(max) = 0

Methods

impl TableComputation

fn new(potential: &PairPotential, size: usize, max: f64) -> TableComputation

Create a new TableComputation for potential, with size points and a maximum value of max.

Examples

use lumol::energy::TableComputation;
use lumol::energy::Harmonic;

let potential = Harmonic{x0: 0.5, k: 4.2};
let table = TableComputation::new(&potential, 1000, 2.0);

assert_eq!(table.energy(1.0), -4.2);
assert_eq!(table.energy(3.0), 0.0);Run