Struct lumol::energy::Wolf

pub struct Wolf { /* fields omitted */ }

Wolf summation for coulombic interactions.

This is a fast, direct, pairwise summation for coulombic potential [Wolf1999].

Examples

use lumol::energy::Wolf;
use lumol::units;

// A relatively large cutoff is needed for Wolf summation
let wolf = Wolf::new(12.0);

use lumol::sys::System;
use lumol::sys::Particle;
use lumol::sys::UnitCell;
use lumol::types::Vector3D;

// Setup a system containing a NaCl pair
let mut system = System::new();
system.set_cell(UnitCell::cubic(10.0));

let mut na = Particle::new("Na");
na.charge = 1.0;
na.position = Vector3D::new(0.0, 0.0, 0.0);

let mut cl = Particle::new("Cl");
cl.charge = -1.0;
cl.position = Vector3D::new(2.0, 0.0, 0.0);

system.add_particle(na);
system.add_particle(cl);

// Use Wolf summation for electrostatic interactions
system.interactions_mut().set_coulomb(Box::new(wolf));

assert_eq!(system.potential_energy(), -0.07292902695393541);Run

[Wolf1999]: Wolf, D. et al. J. Chem. Phys. 110, 8254 (1999).

Methods

impl Wolf

fn new(cutoff: f64) -> Wolf

Create a new Wolf summation, using a real-space cutoff of cutoff.