Trait lumol::energy::GlobalCache

pub trait GlobalCache {
    fn move_particles_cost(&mut self, system: &System, idxes: &[usize], newpos: &[Vector3D]) -> f64;
    fn update(&mut self);
}

Energetic cache for global potentials.

This trait provide all the functions needed by EnergyCache to compute partial energy updates in Monte-Carlo simulations. You can use a panic!ing implementation for all methods if you never need to use a given GlobalPotential in Monte-Carlo simulations.

Examples

use lumol::energy::{GlobalPotential, GlobalCache};
use lumol::types::{Vector3D, Matrix3, Zero};
use lumol::sys::{System, Particle};

/// Shift the energy of all the particles by a given delta.
#[derive(Clone)]
struct ShiftAll {
    delta: f64,
}

impl GlobalPotential for ShiftAll {
    fn energy(&mut self, system: &System) -> f64 {
        // shift all particles by delta
        self.delta * system.size() as f64
    }

    fn forces(&mut self, system: &System) -> Vec<Vector3D> {
        // this potential does not changes the forces
        vec![Vector3D::zero(); system.size()]
    }

    fn virial(&mut self, system: &System) -> Matrix3 {
        // the virial is null as all the forces are null
        Matrix3::zero()
    }
}

impl GlobalCache for ShiftAll {
    fn move_particles_cost(&mut self, _: &System, _: &[usize], _: &[Vector3D]) -> f64 {
        // The cost of moving particles is null, because all the particles
        // get the same energy shift whatever there position are.
        return 0.0
    }

    fn update(&mut self) {
        // We are not storing anything in the ShiftAll struct, so this
        // function is a no-op.
    }
}Run

Required Methods

fn move_particles_cost(&mut self, system: &System, idxes: &[usize], newpos: &[Vector3D]) -> f64

Get the cost of moving particles in the system.

This function is passed the current system, and two slices with the same size: idxes contains the indexes of the moves particles (getting the first moved particle is a call to system[idxes[0]]); and newpos contains the new positions of the particles. The previous positions of the particles are still in the system.

fn update(&mut self)

Update the cache as needed after a call to move_particles_cost.

If the Monte-Carlo move is accepted, this function will be called and should update any cached quantity so that further call to GlobalPotential::energy gives the right value.