Trait lumol::out::Output

pub trait Output {
    fn write(&mut self, system: &System);

    fn setup(&mut self, _: &System) { ... }
    fn finish(&mut self, _: &System) { ... }
}

The Output trait define the interface for all the quantities outputted by the simulation during the run. An Output can be a text or a binary data file, an image, a text log, …

Required Methods

fn write(&mut self, system: &System)

Write the output from the system.

Provided Methods

fn setup(&mut self, _: &System)

Function called once at the beginning of the simulation, which allow for some setup of the output if needed.

fn finish(&mut self, _: &System)

Function called once at the end of the simulation.