Module lumol::sim

Types and traits for representing simulation algorithms

Modules

mc
Monte-Carlo Metropolis algorithms

md
Molecular dynamics algorithms.

min
Energy minimization algorithms

Structs

MolecularDynamics
Molecular Dynamics propagator for the simulation.

MonteCarlo
Metropolis Monte-Carlo propagator

Simulation
The Simulation struct holds all the needed algorithms for running the simulation. It should be use together with a System to perform the simulation.

SteepestDescent
Steepest gradient descent for energy minimization.

Enums

TemperatureStrategy
Possible temperature computation strategies. Different propagators needs different ways to compute the temperature: Monte-Carlo temperature is a constant of the simulation, whereas for molecular dynamics we use the instantaneous velocities.

Traits

Propagator
The propagator trait is the main algorithm of a simulation, i.e. the one which update the system. The main function here is propagate, which should propagate the simulation for one step.