Module lumol::sim::mc

Metropolis Monte Carlo related algorithms

Structs

MonteCarlo
Metropolis Monte Carlo propagator

MonteCarloBuilder
Builder for MonteCarlo struct

Resize
Monte Carlo move that changes the size of the simulation cell

Rotate
Monte Carlo move for rotating a rigid molecule

Translate
Monte Carlo move for translating a molecule

Enums

MCDegreeOfFreedom
Possible degrees of freedom simulated by a given Monte Carlo move

Traits

MCMove
The MCMove trait correspond to the set of methods used in Monte Carlo simulations.