Struct lumol::sim::mc::Rotate

pub struct Rotate { /* fields omitted */ }

Monte-Carlo move for rotating a rigid molecule

Methods

impl Rotate

fn new(theta: f64) -> Rotate

Create a new Rotate move, with maximum angular displacement of theta, rotating all the molecules in the system.

fn with_moltype(theta: f64, moltype: u64) -> Rotate

Create a new Rotate move, with maximum angular displacement of theta, rotating only molecules with moltype type.