Trait lumol::sim::mc::MCMove

pub trait MCMove {
    fn describe(&self) -> &str;
    fn prepare(&mut self, system: &mut System, rng: &mut Box<Rng>) -> bool;
    fn cost(&self, system: &System, beta: f64, cache: &mut EnergyCache) -> f64;
    fn apply(&mut self, system: &mut System);
    fn restore(&mut self, system: &mut System);
}

The MCMove trait correspond to the set of methods used in Monte-Carlo simulations.

Required Methods

fn describe(&self) -> &str

Give a short description of this move

fn prepare(&mut self, system: &mut System, rng: &mut Box<Rng>) -> bool

Prepare the move, by selecting the particles to move, and the parameters of the move. The rng random number generator should be used to generate the parameters of the move.

This function should return true is we can perform the move, and false otherwise.

fn cost(&self, system: &System, beta: f64, cache: &mut EnergyCache) -> f64

Get the cost of performing this move on system. For example in simple NVT simulations, this cost is the energetic difference over beta. The cost must be dimmensionless, and will be placed in an exponential. The cache should be used to compute the cost, or the cache.unused function should be used to ensure that the cache is updated as needed after this move.

fn apply(&mut self, system: &mut System)

Effectivelly apply the move, if it has not already been done in prepare.

fn restore(&mut self, system: &mut System)

Restore the system to it's initial state, if it has been changed in prepare.