Module lumol::sim::md

Molecular dynamics algorithms.

Structs

AnisoBerendsenBarostat
Anisotropic Berendsen barostat integrator based on velocity-Verlet. This one neither reversible nor symplectic.

BerendsenBarostat
Berendsen barostat integrator based on velocity-Verlet. This one neither reversible nor symplectic.

BerendsenThermostat
Berendsen thermostat.

LeapFrog
Leap-frog integrator. This one is reversible and symplectic.

MolecularDynamics
Molecular Dynamics propagator for the simulation.

RemoveRotation
Remove global rotation from the system

RemoveTranslation
Remove global translation from the system

RescaleThermostat
Velocity rescaling thermostat.

Rewrap
Rewrap all molecules' centers of mass to lie within the unit cell. Individual atoms in a molecule may still lie outside of the cell.

VelocityVerlet
Velocity-Verlet integrator. This one is reversible and symplectic.

Verlet
Verlet integrator. This one is reversible and symplectic.

Traits

Control
Trait for controlling some parameters in a system during a simulation.

Integrator
The Integrator trait define integrator interface for molecular dynamics. An integrator is an algorithm responsible for propagating the equations of motion in the system.

Thermostat
Trait for controls usable as thermostats