Module lumol::sim::output

Saving properties of a system during a simulation

Structs

CellOutput
The CellOutput writes all the components of a cell to a file . The columns in the file contain the following values: step A B C α β γ.

CustomOutput
The CustomOutput writes data into a file from an user-provided template.

EnergyOutput
The EnergyOutput writes the energy of the system to a text file, organized as: steps PotentialEnergy KineticEnergy TotalEnergy.

ForcesOutput
The ForcesOutput writes the forces acting on the atoms using XYZ format

PropertiesOutput
The PropertiesOutput write various physical properties of the system to a file. These properties are:

StressOutput
The StressOutput writes the stress of the system to a text file, organized as: step stress.xx stress.yy stress.zz stress.xy stress.xz stress.yz.

TrajectoryOutput
The TrajectoryOutput allows to write the trajectory of the system to a file, using any format supported by the Chemfiles library.

Enums

CustomOutputError
Possible causes of error when using a custom output

Traits

Output
The Output trait defines the interface for all the quantities outputted by the simulation during the run. An Output can be a text or a binary data file, an image, a text log, …