Trait lumol::sim::output::Output

pub trait Output {
    fn write(&mut self, system: &System);

    fn setup(&mut self, &System) { ... }

    fn finish(&mut self, &System) { ... }
}

The Output trait defines the interface for all the quantities outputted by the simulation during the run. An Output can be a text or a binary data file, an image, a text log, …

Required methods

fn write(&mut self, system: &System)

Write the output from the system.

Loading content...
Provided methods

fn setup(&mut self, &System)

Function called once at the beginning of the simulation, which allows for some setup of the output if needed.

fn finish(&mut self, &System)

Function called once at the end of the simulation.

Loading content...