Struct lumol::sim::Simulation

pub struct Simulation { /* fields omitted */ }

The Simulation struct holds all the needed algorithms for running the simulation. It should be use together with a System to perform the simulation.

Methods

impl Simulation

fn new(propagator: Box<Propagator>) -> Simulation

Create a new Simulation from a Propagator.

fn run(&mut self, system: &mut System, nsteps: usize)

Run the simulation on System for nsteps steps.

fn add_output(&mut self, output: Box<Output>)

Add a new Output algorithm in the outputs list

fn add_output_with_frequency(&mut self, output: Box<Output>, frequency: u64)

Add a new Output algorithm in the outputs list, which will be used at the given frequency. The output will be used every time the system step matches this frequency.