Trait lumol::sim::Propagator

pub trait Propagator {
    fn temperature_strategy(&self) -> TemperatureStrategy;
    fn propagate(&mut self, system: &mut System);

    fn setup(&mut self, _: &System) { ... }
    fn finish(&mut self, _: &System) { ... }
}

The propagator trait is the main algorithm of a simulation, i.e. the one which update the system. The main function here is propagate, which should propagate the simulation for one step.

Required Methods

fn temperature_strategy(&self) -> TemperatureStrategy

Get the temperature computation strategy for this propagator

fn propagate(&mut self, system: &mut System)

Propagate the system for one simulation step.

Provided Methods

fn setup(&mut self, _: &System)

Setup code, preparing all the meta-information needed about the simulation.

fn finish(&mut self, _: &System)

Finish the simulation, and maybe output some information about it