Module lumol::sys

The system module provide a way to store data about a simulated system.

Modules

compute
Computing properties of a system

veloc
This module provides some ways to initialize the velocities in a System

Structs

Angle
An Angle formed by the particles at indexes i, j and k

Bond
A Bond between the particles at indexes i and j

BondDistance
The BondDistance bitflag encode the topological distance between two particles in the molecule, i.e. the number of bonds between the particles.

Composition
The system composition contains the number of particles of each kind in the system.

Configuration
The Configuration contains the physical data of the system:

Dihedral
A Dihedral angle formed by the particles at indexes i, j, k and m

ElementData
Data about one "extended" chemical element.

EnergyCache
This is a cache for energy computation.

EnergyEvaluator
An helper struct to evaluate energy components of a system.

Molecule
A molecule is the basic building block for a topology. It contains data about the connectivity (bonds, angles, dihedrals) in the system.

Particle
The Particle type hold basic data about a particle in the system. It is self contained, so that it will be easy to send data between parallels processes.

ParticleKind
A particle kind. Particles with the same name will have the same kind. This is used for faster potential lookup.

PeriodicTable
The PeriodicTable struct give access to elements information, and give a way to register new elements in the list.

System
The System type hold all the data about a simulated system.

Trajectory
A Trajectory is a file containing one or more successive simulation steps.

TrajectoryBuilder
A Trajectory builder, to set some options before opening a trajectory.

TrajectoryError
Possible error when reading and writing to trajectories

UnitCell
An UnitCell defines the system physical boundaries.

Enums

CellShape
The shape of a cell determine how we will be able to compute the periodic boundaries condition.

OpenMode
Possible modes when opening a Trajectory.

Constants

BONDED_12
The particles are separated by one bond

BONDED_13
The particles are separated by two bonds

BONDED_14
The particles are separated by three bonds

BONDED_FAR
The particles are separated by more than three bonds

Functions

molecule_type
Get the molecule type of the given molecule containing the particles. This type can be used to identify all the molecules containing the same bonds and particles (see System::molecule_type for more information).

read_molecule
Read a the first molecule from the file at path. If no bond information exists in the file, bonds are guessed.