Module lumol::sys::compute

Algorithm to compute physical properties of a System

Structs

AtomicVirial
Compute the virial tensor of the system using the atomic definition.

Forces
Compute all the forces acting on the system, and return a vector of force acting on each particles

KineticEnergy
Compute the kinetic energy of the system

MolecularVirial
Compute the virial tensor of the system using the molecular definition

PotentialEnergy
Compute the potential energy of the system

Pressure
Compute the pressure of the system using the virial definition.

PressureAtTemperature
Compute the pressure of the system using the virial definition, at a given temperature.

Stress
Compute the stress tensor of the system from the virial definition

StressAtTemperature
Compute the stress tensor of the system from the virial definition, at the given temperature.

Temperature
Compute the instantaneous temperature of the system

TotalEnergy
Compute the total energy of the system

Virial
Compute the virial tensor of the system, picking between AtomicVirial and MolecularVirial depending on the number of degrees of freedom simulated on the system.

Volume
Compute the volume of the system

Traits

Compute
The Compute trait allow to compute properties of a system, without modifying this system. The Output type is the type of the computed property.