Struct lumol::sys::Configuration

pub struct Configuration {
    pub cell: UnitCell,
    // some fields omitted
}

The Configuration contains the physical data of the system:

The unit cell;

The list of particles in the system;

The list of molecules in the system.

Fields
cell: UnitCell

Unit cell of the system

Methods

impl Configuration

fn new() -> Configuration

Create a new empty Configuration

impl Configuration

Topology and particles related functions

fn molecule_type(&self, molid: usize) -> u64

Get the type of the molecule at index molid. This type is a hash of the atoms names, and the set of bonds in the molecule. This means that two molecules will have the same type if and only if they contains the same atoms and the same bonds, in the same order.

fn molecules_with_moltype(&self, moltype: u64) -> Vec<usize>

Get a list of molecules with moltype molecule type.

fn are_in_same_molecule(&self, i: usize, j: usize) -> bool

Check if the particles at indexes i and j are in the same molecule

fn molecules(&self) -> &[Molecule]

Get the list of molecules in the configuration.

fn molecule(&self, id: usize) -> &Molecule

Get the molecule at index id

fn molid(&self, i: usize) -> usize

Get the index of the molecule containing the particle i

fn bond_distance(&self, i: usize, j: usize) -> i8

Get the length of the shortest bond path to go from the particle i to the particle j. If the particles are not in the same molecule, the length is -1. Else, this length is 0 if i == j, 1 if there is a bond between i and j, etc.

fn remove_molecule(&mut self, molid: usize)

Remove the molecule at index i

fn add_bond(
&mut self,
particle_i: usize,
particle_j: usize
) -> Vec<(usize, usize)>

Add a bond between the particles at indexes i and j. The particles should have been added to the configuration before calling this.

Warning

If the bond is between two particles which are not in the same molecule, the two molecules are merged together by moving particles in the particles list, and thus invalidate any previously stored index. In particular, any bond, angle, dihedral or molecule is invalidated.

This function will return the list of atomic permutations that where applied in order to ensure that molecules are contiguous in memory.

fn remove_particle(&mut self, i: usize)

Removes particle at index i and any associated bonds, angle or dihedral

fn add_particle(&mut self, particle: Particle)

Insert a particle at the end of the internal list. The new particle must have a valid particle kind.

fn size(&self) -> usize

Get the number of particles in this configuration

fn is_empty(&self) -> bool

Check if this configuration contains any particle

fn molecule_com(&self, molid: usize) -> Vector3D

Return the center-of-mass of a molecule

Warning

This function does not check for the particles' positions' nearest images. To use this function properly, make sure that all particles of the molecule are adjacent.

fn center_of_mass(&self) -> Vector3D

Return the center-of-mass of the configuration

Warning

This function does not check for the particles' positions' nearest images. To use this function properly, make sure that all particles of a molecule are adjacent.

fn wrap_molecule(&mut self, molid: usize)

Move all particles of a molecule such that the molecules center-of-mass position resides inside the simulation cell.

Note

If the CellShape is Infinite there are no changes to the positions.

fn particles(&self) -> Iter<Particle>

Get an iterator over the Particle in this configuration

fn particles_mut(&mut self) -> IterMut<Particle>

Get a mutable iterator over the Particle in this configuration

fn particle(&self, i: usize) -> &Particle

Get a reference to the i'th particle in this configuration

fn particle_mut(&mut self, i: usize) -> &mut Particle

Get a mutable reference to the i'th particle in this configuration

impl Configuration

UnitCell related functions

fn distance(&self, i: usize, j: usize) -> f64

Get the distance between the particles at indexes i and j

fn nearest_image(&self, i: usize, j: usize) -> Vector3D

Get the vector between the nearest image of particle j with respect to particle i.

fn angle(&self, i: usize, j: usize, k: usize) -> f64

Get the angle between the particles i, j and k

fn angle_and_derivatives(
&self,
i: usize,
j: usize,
k: usize
) -> (f64, Vector3D, Vector3D, Vector3D)

Get the angle and the derivatives of the angle between the particles i, j and k

fn dihedral(&self, i: usize, j: usize, k: usize, m: usize) -> f64

Get the dihedral angle between the particles i, j, k and m

fn dihedral_and_derivatives(
&self,
i: usize,
j: usize,
k: usize,
m: usize
) -> (f64, Vector3D, Vector3D, Vector3D, Vector3D)

Get the dihedral angle and the derivatives of the dihedral angle between the particles i, j, k and m