Struct lumol::sys::ParticleSlice

pub struct ParticleSlice<'a> {
    pub name: &'a [String],
    pub kind: &'a [ParticleKind],
    pub charge: &'a [f64],
    pub mass: &'a [f64],
    pub position: &'a [Vector3D],
    pub velocity: &'a [Vector3D],
}

A slice of Particle inside a ParticleVec .

Fields
name: &'a [String]

A slice of name from a ParticleVec

kind: &'a [ParticleKind]

A slice of kind from a ParticleVec

charge: &'a [f64]

A slice of charge from a ParticleVec

mass: &'a [f64]

A slice of mass from a ParticleVec

position: &'a [Vector3D]

A slice of position from a ParticleVec

velocity: &'a [Vector3D]

A slice of velocity from a ParticleVec

Methods

impl<'a> ParticleSlice<'a>

pub fn len(&self) -> usize

Similar to [Particle] ::len(), the length of all fields should be the same.

pub fn is_empty(&self) -> bool

Similar to [Particle] ::is_empty(), the length of all fields should be the same.

pub fn first(&self) -> Option<ParticleRef<'a>>

Similar to [Particle] ::first().

pub fn split_first(&self) -> Option<(ParticleRef<'a>, ParticleSlice<'a>)>

Similar to [Particle] ::split_first().

pub fn last(&self) -> Option<ParticleRef<'a>>

Similar to [Particle] ::last().

pub fn split_last(&self) -> Option<(ParticleRef<'a>, ParticleSlice<'a>)>

Similar to [Particle] ::split_last().

pub fn split_at(&self, mid: usize) -> (ParticleSlice<'a>, ParticleSlice<'a>)

Similar to [Particle] ::split_at().

pub fn get(&self, i: usize) -> Option<ParticleRef>

Similar to [Particle] ::get().

pub unsafe fn get_unchecked(&self, i: usize) -> ParticleRef

Similar to [Particle] ::get_unchecked().

pub fn as_ptr(&self) -> ParticlePtr

Similar to [Particle] ::as_ptr().

pub unsafe fn from_raw_parts<'b>(
data: ParticlePtr,
len: usize
) -> ParticleSlice<'b>

Similar to std::slice::from_raw_parts().

impl<'a> ParticleSlice<'a>

pub fn to_vec(&self) -> ParticleVec

Similar to [Particle] ::to_vec().

impl<'a> ParticleSlice<'a>

pub fn iter(&self) -> Iter

Get an iterator over the ParticleRef in this slice.