Struct lumol::sys::System

pub struct System {
    pub simulated_degrees_of_freedom: DegreesOfFreedom,
    pub step: u64,
    // some fields omitted
}

The System type hold all the data about a simulated system.

This data contains:

an unit cell, containing the system;

a list of particles in the system;

a list of molecules in the system;

a list of interactions, associating particles kinds and potentials

In the implementation, the particles contained in a molecule are guaranteed to be contiguous in memory. This allow for faster access when iterating over molecules, and easier molecule removal from the system.

Fields
simulated_degrees_of_freedom: DegreesOfFreedom

Number of degrees of freedom simulated in the system. This default to DegreesOfFreedom::Particles, and is set in the simulation setup.

step: u64

The current simulation step

Methods

impl System

pub fn new() -> System

Create a new empty System

pub fn with_cell(cell: UnitCell) -> System

Create an empty system with a specific unit cell

pub fn add_molecule(&mut self, molecule: Molecule)

Add a molecule to the system

pub fn composition(&self) -> Composition

Get the composition in particles and molecules of the configuration

pub fn simulated_temperature(&mut self, temperature: Option<f64>)

Use an external temperature for all the system properties. Calling this with Some(temperature) will replace all the computation of the temperature from the velocities with the given values. Calling it with None will use the velocities.

The default is to use the velocities unless this function is called.

impl System

Functions related to interactions

pub fn energy_evaluator(&self) -> EnergyEvaluator

Get an helper struct to evaluate the energy of this system.

pub fn set_pair_potential(&mut self, (&str, &str), potential: PairInteraction)

Set the pair interaction potential for atoms with types i and j

pub fn set_bond_potential(
&mut self,
(&str, &str),
potential: Box<dyn BondPotential + 'static>
)

Set the bond interaction potential for atoms with types i and j

pub fn set_angle_potential(
&mut self,
(&str, &str, &str),
potential: Box<dyn AnglePotential + 'static>
)

Set the angle interaction potential for atoms with types i, j, and k

pub fn set_dihedral_potential(
&mut self,
(&str, &str, &str, &str),
potential: Box<dyn DihedralPotential + 'static>
)

Set the dihedral angle interaction potential for atoms with types i, j, k, and m.

pub fn set_coulomb_potential(
&mut self,
potential: Box<dyn CoulombicPotential + 'static>
)

Set the coulombic interaction for all pairs to potential

pub fn add_global_potential(
&mut self,
potential: Box<dyn GlobalPotential + 'static>
)

Add the potential global interaction

pub fn pair_potential(&self, i: usize, j: usize) -> Option<&PairInteraction>

Get the pair potential acting between the particles at indexes i and j.

pub fn bond_potential(&self, i: usize, j: usize) -> Option<&dyn BondPotential>

Get the bond potential acting between the particles at indexes i and j.

pub fn angle_potential(
&self,
i: usize,
j: usize,
k: usize
) -> Option<&dyn AnglePotential>

Get the angle potential acting between the particles at indexes i, j and k.

pub fn dihedral_potential(
&self,
i: usize,
j: usize,
k: usize,
m: usize
) -> Option<&dyn DihedralPotential>

Get the dihedral angles potential acting between the particles at indexes i, j, k and m.

pub fn coulomb_potential(&self) -> Option<&dyn CoulombicPotential>

Get the coulombic interaction for the system

pub fn global_potentials(&self) -> &[Box<dyn GlobalPotential + 'static>]

Get all global interactions for the system

pub fn maximum_cutoff(&self) -> Option<f64>

Get maximum cutoff from coulomb, pairs and global interactions.

impl System

Functions to get physical properties of a system.

pub fn degrees_of_freedom(&self) -> usize

Get the number of degrees of freedom in the system

pub fn kinetic_energy(&self) -> f64

Get the kinetic energy of the system.

pub fn potential_energy(&self) -> f64

Get the potential energy of the system.

pub fn total_energy(&self) -> f64

Get the total energy of the system.

pub fn temperature(&self) -> f64

Get the temperature of the system.

pub fn volume(&self) -> f64

Get the volume of the system.

pub fn virial(&self) -> Matrix3

Get the virial of the system as a tensor

pub fn pressure(&self) -> f64

Get the pressure of the system from the virial equation, at the system instantaneous temperature.

pub fn stress(&self) -> Matrix3

Get the stress tensor of the system from the virial equation.

pub fn forces(&self) -> Vec<Vector3D>

Get the forces acting on all the particles in the system

impl System

pub fn check(&self)

Check the system before running a simulation