Struct lumol::sys::Trajectory

pub struct Trajectory(_);

A Trajectory is a file containing one or more successive simulation steps.

One should use the TrajectoryBuilder to create a new trajectory.

Examples

let mut trajectory = TrajectoryBuilder::new()
                                      .open("file.xyz")
                                      .unwrap();

let system = trajectory.read().unwrap();

Methods

impl Trajectory

fn read(&mut self) -> Result<System, TrajectoryError>

Read the next step of the trajectory

Examples

let mut trajectory = TrajectoryBuilder::new()
                                      .open("file.nc")
                                      .unwrap();

let system = trajectory.read().unwrap();

fn read_guess_bonds(&mut self) -> Result<System, TrajectoryError>

Read the next step of the trajectory, and guess the bonds of the resulting [System][struct.System.html].

Examples

let mut trajectory = TrajectoryBuilder::new()
                                      .open("file.nc")
                                      .unwrap();

let system = trajectory.read_guess_bonds().unwrap();

fn write(&mut self, system: &System) -> Result<(), TrajectoryError>

Write the system to the trajectory.

Examples

let mut trajectory = TrajectoryBuilder::new()
                                      .mode(OpenMode::Write)
                                      .open("file.xyz")
                                      .unwrap();

trajectory.write(&system).unwrap();

fn set_cell(&mut self, cell: &UnitCell) -> Result<(), TrajectoryError>

Set the unit cell associated with a trajectory. This cell will be used when reading and writing the files, replacing any unit cell in the frames or files.

Examples

let mut trajectory = TrajectoryBuilder::new()
                                      .open("file.xyz")
                                      .unwrap();

trajectory.set_cell(&UnitCell::cubic(10.0)).unwrap();
let system = trajectory.read().unwrap();

assert_eq!(system.cell, UnitCell::cubic(10.0));

fn set_topology_file(&mut self, path: &str) -> Result<(), TrajectoryError>

Set the topology associated with this trajectory by reading the first frame of the file at the given path and extracting the topology of this frame. This topology will be used to replace any existing topology when reading or writing with this trajectory.

Examples

let mut trajectory = TrajectoryBuilder::new()
                                      .open("file.xyz")
                                      .unwrap();

trajectory.set_topology_file("topology.pdb").unwrap();
// The system will contain the topology from topology.pdb
let system = trajectory.read().unwrap();