Struct lumol::types::Array3

pub struct Array3<T>(_);

Three dimensional tensors, based on ndarray

Most of the methods are simply forwarded to ndarray, so also look the doc for this crate. This array type mainly supports indexing using tuples as indices and is though as storage backend for multi-dimensional data.

let mut a = Array3::zeros((3, 5, 2));

assert_eq!(a[(0, 4, 1)], 0.0);

a[(0, 4, 1)] = 7.0;
assert_eq!(a[(0, 4, 1)], 7.0);

Methods

impl<T> Array3<T>

fn zeros(size: (usize, usize, usize)) -> Array3<T> where
T: Zero + Clone,

Create a new Array3 of the specified size filled with the Zero::zero return value.

Examples

let a: Array3<f64> = Array3::zeros((8, 5, 2));
assert_eq!(a[(6, 2, 0)], 0.0);

fn resize_if_different(&mut self, size: (usize, usize, usize)) where
T: Zero + Clone,

Resize the array if the current size is not size, and fill the new array with zeros.

Examples

let mut a = Array3::zeros((8, 5, 7));

a[(3, 3, 3)] = 42.0;

// This does nothing
a.resize_if_different((8, 5, 7));
assert_eq!(a[(3, 3, 3)], 42.0);

// This allocates a new array
a.resize_if_different((8, 5, 6));
assert_eq!(a[(3, 3, 3)], 0.0);

fn default(size: (usize, usize, usize)) -> Array3<T> where
T: Default,

Create a new Array3 of the specified size filled with the Default::default return value. Default::default return value.

Examples

let a: Array3<f64> = Array3::zeros((8, 5, 2));
let b: Array3<f64> = Array3::default((8, 5, 2));

assert_eq!(a, b);